refactor:user files into /user and your files into /user/files
