+++ client/src/components/user/UserSwipeTab.js
@@ -6,7 +6,7 @@ import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 import Typography from '@material-ui/core/Typography';
 import Box from '@material-ui/core/Box';
-import './SwipeTab.css'
+import '../SwipeTab.css'
 import Skeleton from '@material-ui/lab/Skeleton';
 
 const UserPosts = React.lazy(()=>import('./UserPosts'))
